Add transaction hashes to events
