+++ indexer/httpd/src/graphql.rs
@@ -3,6 +3,7 @@ use {
     async_graphql::{Schema, dataloader::DataLoader, extensions},
     indexer_sql::dataloaders::{
         block_events::BlockEventsDataLoader, block_transactions::BlockTransactionsDataLoader,
+        event_transaction::EventTransactionDataLoader,
         transaction_events::TransactionEventsDataLoader,
         transaction_grug::FileTransactionDataLoader,
         transaction_messages::TransactionMessagesDataLoader,
@@ -28,6 +29,13 @@ pub fn build_schema(app_ctx: Context) -> AppSchema {
 
     let block_events_loader = DataLoader::new(
         BlockEventsDataLoader {
+            db: app_ctx.db.clone(),
+        },
+        tokio::spawn,
+    );
+
+    let event_transaction_loader = DataLoader::new(
+        EventTransactionDataLoader {
             db: app_ctx.db.clone(),
         },
         tokio::spawn,
@@ -68,5 +76,6 @@ pub fn build_schema(app_ctx: Context) -> AppSchema {
     .data(transaction_messages_loader)
     .data(transaction_events_loader)
     .data(file_transaction_loader)
+    .data(event_transaction_loader)
     .finish()
 }
