+++ indexer/sql/src/dataloaders.rs
@@ -1,5 +1,6 @@
 pub mod block_events;
 pub mod block_transactions;
+pub mod event_transaction;
 pub mod transaction_events;
 pub mod transaction_grug;
 pub mod transaction_messages;
