+++ indexer/testing/tests/graphql/main.rs
@@ -1,0 +1,1 @@
+pub mod event;
